+++ src/FakeApplication.php
@@ -10,8 +10,6 @@ namespace BlueMvc\Fakes;
 
 use BlueMvc\Core\Base\AbstractApplication;
 use BlueMvc\Core\Exceptions\InvalidFilePathException;
-use BlueMvc\Core\Interfaces\Collections\SessionItemCollectionInterface;
-use BlueMvc\Fakes\Collections\FakeSessionItemCollection;
 use DataTypes\Exceptions\FilePathInvalidArgumentException;
 use DataTypes\FilePath;
 
@@ -42,7 +40,7 @@ class FakeApplication extends AbstractApplication
             $documentRoot .= DIRECTORY_SEPARATOR;
         }
 
-        parent::__construct(FilePath::parse($documentRoot), new FakeSessionItemCollection());
+        parent::__construct(FilePath::parse($documentRoot));
     }
 
     /**
@@ -55,17 +53,5 @@ class FakeApplication extends AbstractApplication
     public function setDebug(bool $isDebug): void
     {
         parent::setDebug($isDebug);
-    }
-
-    /**
-     * Sets the session items.
-     *
-     * @since 1.0.0
-     *
-     * @param SessionItemCollectionInterface $sessionItems The session items.
-     */
-    public function setSessionItems(SessionItemCollectionInterface $sessionItems): void
-    {
-        parent::setSessionItems($sessionItems);
     }
 }