+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Unreleased
+- **Backward compatibility break**: Moved session handling from FakeApplication to FakeRequest.
 - Updated PHP requirement to 7.1.
 
 ## [1.1.0] - 2018-04-29